tests/wpt/tests/css/selectors/i18n/css3-selectors-lang-047.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>[lang="es"], lang="mx-es"</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/css3-selectors/#attribute-selectors'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style type='text/css'>
.test div { width: 50px; }
#box[lang='es'] { width: 100px; }
</style>
</head>
<body>



<div class="test"><div id="box" lang="mx-es">&#xA0;</div></div>


<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 50);
}, "When the lang= value uses a single subtag, it will NOT match against an attribute value where it appears in a different position.");
</script>

<div id='log'></div>

</body>
</html>
